requirements.txt
anyio==3.6.2
ariadne==0.17.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
ecdsa==0.18.0
Flask==2.2.2
flask-oidc==1.4.0
Flask-SQLAlchemy==3.0.2
graphql-core==3.2.3
greenlet==2.0.1
httplib2==0.21.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==3.0.9
python-jose==3.3.0
python-keycloak==2.8.0
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.23.1
stripe==5.0.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
